requirement.txt
acres==0.5.0
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
distro==1.9.0
etelemetry==0.3.1
faiss-cpu==1.12.0
fastapi==0.119.0
filelock==3.20.0
frontend==0.0.3
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.4
openai==2.5.0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pdfminer.six==20250506
pillow==12.0.0
prov==2.1.1
puremagic==1.30
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydot==4.0.1
PyMuPDF==1.26.5
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==7.2.1
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tools==1.0.6
torch==2.9.0
tqdm==4.67.1
traits==7.0.2
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.1
websockets==15.0.1
